* Merged with default branch

+++ android/make_install_package.sh
@@ -3,8 +3,8 @@
 THIS_DIR=$PWD
 SCRIPT_DIR=`dirname $0`
 ANDROID_SRC_DIR=
-PHOTOSHARE_SRC_DIR=$HOME/Documents/workspace/hg/android/PhotoShare
-HAGGLE_VER=0.2
+HAGGLE_SRC_DIR=`basename $THIS_DIR`
+HAGGLE_VER="x.y"
 
 function usage() {
     echo "Usage: $0 ANDROID_SRC_DIR [ PHOTOSHARE_SRC_DIR ]"
@@ -23,8 +23,23 @@ if [ ! -z $2 ]; then
     echo "Looking for PhotoShare in $PHOTOSHARE_DIR"
 fi
 
+if [ ! -f $HAGGLE_SRC_DIR/configure.ac ]; then
+    if [ -f $ANDROID_SRC_DIR/external/haggle/configure.ac ]; then
+	HAGGLE_SRC_DIR=$ANDROID_SRC_DIR/external/haggle
+    fi
+fi
 
-ANDROID_BIN_DIR=$ANDROID_SRC_DIR/out/target/product/dream-open/system
+if [ ! -f $HAGGLE_SRC_DIR/configure.ac ]; then
+    echo "Could not find Haggle source directory"
+    return;
+fi
+
+HAGGLE_VER=`awk '/AC_INIT/ { l=match($2, /[0-9]/); r=match($2,/\]/); print substr($2,l,r-l) }' $HAGGLE_SRC_DIR/configure.ac`	
+
+PHOTOSHARE_SRC_DIR=$ANDROID_SRC_DIR/external/haggle/android/PhotoShare
+ANDROID_BIN_DIR=$ANDROID_SRC_DIR/out/target/product/passion/system
+IWCONFIG_BIN=$ANDROID_BIN_DIR/xbin/iwconfig
+IWLIST_BIN=$ANDROID_BIN_DIR/xbin/iwlist
 HAGGLE_BIN=$ANDROID_BIN_DIR/bin/haggle
 LIBHAGGLE_SO=$ANDROID_BIN_DIR/lib/libhaggle.so
 LIBHAGGLE_XML2_SO=$ANDROID_BIN_DIR/lib/libhaggle-xml2.so
@@ -47,14 +62,25 @@ mkdir $PACKAGE_DIR
 pushd $PACKAGE_DIR
 
 # Copy all the files we need into our package directory
-for f in "$HAGGLE_BIN $LIBHAGGLE_SO $LIBHAGGLE_XML2_SO $LIBHAGGLE_JNI_SO $HAGGLE_JAR $PHOTOSHARE_APK"; do
-    cp -f $f .
+for f in $HAGGLE_BIN $LIBHAGGLE_SO $LIBHAGGLE_XML2_SO $LIBHAGGLE_JNI_SO $HAGGLE_JAR $PHOTOSHARE_APK $IWCONFIG_BIN $IWLIST_BIN; do
+    echo "Including $f in package"
+
+    if [ -f $f ]; then
+	cp -f $f .
+    fi
 done
 
 # Create install script
 
 cat > install.sh <<EOF
 #!/bin/bash
+
+if [ -z `which adb` ]; then
+    echo "Unable to find the adb command (Android Debug Bridge)."
+    echo "You need to install and configure the Android SDK."
+    echo "See http://developer.android.com"
+    return;
+fi
 
 echo "This script will install Haggle and PhotoShare on any Android devices connected to this computer."
 echo "Note that the installation will only be successful if you have an Android developer phone,"
@@ -83,30 +109,50 @@ adb root
 for dev in \$DEVICES; do
     echo "Installing onto device \$dev"
     # Remount /system partition in read/write mode
-    adb -s \$dev shell su -c mount -o remount,rw -t yaffs2 /dev/block/mtdblock3 /system
-    adb -s \$dev push haggle /system/bin/haggle
-    adb -s \$dev shell su -c chmod 4775 /system/bin/haggle
+    adb -s \$dev shell mount -o remount,rw -t yaffs2 /dev/block/mtdblock3 /system
 
+    echo "Installing Haggle binary..."
+    adb -s \$dev push haggle /system/bin/haggle
+    adb -s \$dev shell chmod 4775 /system/bin/haggle
+
+    echo "Installing libraries..."
     adb -s \$dev push libhaggle.so /system/lib/libhaggle.so
-    adb -s \$dev shell su -c chmod 644 /system/lib/libhaggle.so
+    adb -s \$dev shell chmod 644 /system/lib/libhaggle.so
 
     adb -s \$dev push libhaggle-xml2.so /system/lib/libhaggle-xml2.so
-    adb -s \$dev shell su -c chmod 644 /system/lib/libhaggle-xml2.so
+    adb -s \$dev shell chmod 644 /system/lib/libhaggle-xml2.so
 
     adb -s \$dev push libhaggle_jni.so /system/lib/libhaggle_jni.so
-    adb -s \$dev shell su -c chmod 644 /system/lib/libhaggle_jni.so
+    adb -s \$dev shell chmod 644 /system/lib/libhaggle_jni.so
 
+    echo "Installing haggle.jar"
     adb -s \$dev push haggle.jar /system/framework/haggle.jar
-    adb -s \$dev shell su -c chmod 644 /system/framework/haggle.jar
+    adb -s \$dev shell chmod 644 /system/framework/haggle.jar
 
+    echo "Installing tools..."
+    if [ -f iwconfig ]; then
+        adb -s \$dev push iwconfig /system/bin/iwconfig
+        adb -s \$dev shell chmod 4775 /system/bin/iwconfig
+    fi
+    if [ -f iwlist ]; then
+        adb -s \$dev push iwlist /system/bin/iwlist
+        adb -s \$dev shell chmod 4775 /system/bin/iwlist
+    fi
+
+    echo "Uninstalling PhotoShare application..."
     adb -s \$dev uninstall org.haggle.PhotoShare
+    echo "Installing PhotoShare application..."
     adb -s \$dev install PhotoShare.apk
 
     # Reset /system partition to read-only mode
-    adb -s \$dev shell su -c mount -o remount,ro -t yaffs2 /dev/block/mtdblock3 /system
+    adb -s \$dev shell mount -o remount,ro -t yaffs2 /dev/block/mtdblock3 /system
+
+    # Create data folder if it does not exist already
+    adb -s \$dev shell mkdir /data/haggle &>/dev/null
 
     # Cleanup data folder if any
-    adb -s \$dev shell su -c rm /data/haggle/* &> /dev/null
+    adb -s \$dev shell rm /data/haggle/* &> /dev/null
+    
 done
 
 EOF
@@ -115,7 +161,8 @@ chmod +x install.sh
 
 cat > README <<EOF
 You need an Android Developer phone, or other Android phone with root access,
-in order to install Haggle.
+in order to install Haggle. You also need to install the Android SDK so that
+you have the Android debug bridge (adb) tool.
 
 To install, connect your Android phone to a computer and run ./install.sh
